Add a dagify_from algorithm, with a somewhat cheating implementation
